requirements.txt
boto3==1.34.70
python-telegram-bot==21.1.1